normal mapping
lesson about normal maps, which add depth to our textures will little extra costs, we use a normal map (0.5,0.5,1 of colors) and normalize to clear effect.
to make effect rotation safe, we transform the light and other fragment stuff into tangent space (tbn) and after doing the calcuations in the vertex we send it to the fragment bunched together.