hdr
hdr lesson, brightness in most screens is limited to a range of 0-1, we can bypass it by compress it to our normal range using a framebuffer wiht gamma correction. (tonoe mapping).
we use colors 16f/32f in our framebuffer to register that.
we than use var exposure to set the brightness of  the light