deferred shading
very interesting lesson about rendering techniques: forward vs deferred
in deferred unlike how we did so fard in forward, we use something called a gbuffer which groups all geometry, lighting methods, positions, and normal maps/depth maps together.
it is more efficient for bigger projects but comes with the cost that stuff like msaa is not possible, but theres workarounds, so usually the answer isnt black and white, and we use a mix or defereed with workarounds.